Chart Working With API
