--- lib/Models/CoinMarketAPI/market_fetch_coins_models.dart
@@ -1,5 +1,0 @@
-export 'market_usd_model.dart';
-export 'market_data_model.dart';
-export 'market_quote_model.dart';
-export 'market_status_model.dart';
-export 'market_big_data_model.dart';
