+++ lib/Models/MessariAPI/fetch_coins_models.dart
@@ -1,5 +1,2 @@
-export 'usd_model.dart';
-
-export 'quote_model.dart';
 export 'status_model.dart';
 export 'big_data_model.dart';
