+++ lib/first_screen.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'Models/Chart.dart';
 import 'Models/MessariAPI/big_data_model.dart';
 import 'Models/MessariAPI/Repository.dart';
 import 'Models/DetailsPageModel.dart';
